requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
flake8==7.3.0
gunicorn==23.0.0
inflection==0.5.1
isort==6.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
packaging==25.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
whitenoise==6.9.0
